refactor(put_file): migrate method rdm_add_file to requests_rdm
Migrate method `rdm_add_file` to `requests_rdm`, since it is
semantically more coherent in that module and there is no
need for it to be in a separate module, only containing this
single method.
